refactor(timeline): split clip lane tests into their own file
clips.rs had grown to 1,210 lines, past the project's 1,000-line
ceiling, and the box-select work lands more code in it. Move the test
module to clips_tests.rs; no test bodies change.
